tty: serial: fsl_lpuart: Fix for Tx DMA when circ_buf head is 0
When transmitting from a circular buffer, the fsl_lpuart driver creates
a scatter-gather list for DMA tranmit if head < tail on the uart
circular buffer.  However, if head is exactly 0 only one is needed, and
the second 0 length sg was causing a DMA error that the uart did not
recover from.